+++ CHANGELOG.md
@@ -47,7 +47,7 @@
 ## 2.0.2 (Nov 6, 2023)
 
 - Changed bootstrap order to allow overriding favicon options in `config/web.php`
-- Moved `AdminButton` class to `Hirtz\Skeleton\widgets\AdminButton`
+- Moved `AdminButton` class to `Hirtz\Skeleton\Widgets\AdminButton`
 
 ## 2.0.1 (Nov 5, 2023)